Ks/trying to merge (#346)
* added server into src

* reconfigured package.jsons

* added a time filter

* trying to merge
