+++ package.json
@@ -3,7 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "concurrently \"npm run dev:next\" \"npm run dev:server\"",
+    "dev:next": "next dev",
+    "dev:server": "node src/server/server.js",
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
@@ -13,7 +15,15 @@
     "prepare": "husky"
   },
   "dependencies": {
+    "@notionhq/client": "^2.2.15",
+    "body-parser": "^1.20.3",
     "bootstrap": "^5.3.2",
+    "concurrently": "^9.1.0",
+    "cors": "^2.8.5",
+    "dotenv": "^16.4.5",
+    "express": "^4.21.2",
+    "find-config": "^1.0.0",
+    "moment": "^2.30.1",
     "next": "^14.2.10",
     "react": "^18",
     "react-bootstrap": "^2.9.0",
